Player paddle movement
Implementation of the player paddle input. The player can now move the
paddle up and down using the arrow keys.
Changed Canvas to a JPanel as previoussly unable to add components to
the componenet tree through the Canavas, making it near impossible to
handle inputs. Also mixing AWT and Spring components is considered Bad.
The Arena now extends a JPanel which in turns provides increased
performance optimisations over the Canvas.