Timer changes and arena graphics
Changed the game timer (game loop) to a swing timer as the Java.util.Timer was not acting as expected. There was rendering issues and some input lag. Also with this changed the loop so that it exectues game updates first then renders.

Also added some inital arena graphics, i.e. black background and a center line.