Ability for both the Player and Computer to score
Both the Player and Computer entities can now score a point in the game,
up to a maximum of 6 (currently), which then tirggers the end of the
game.
The system is implemented using a observe/notify system, whereby the
Ball handles the checking of its position (i.e. in the score zone) and
notifies all relevant listeners to update the score value.

The Arena now checks the end of the game scenario (i.e. the Player or
the Computer hits the max score), and displays the current score of the
2 Players