Smooth Player movement
Smooth out player movement by tracking the delta which is changed on
input. Pressing the control increases the delta and releasing the
control returns it to normal. The interpolation of this move is now
carried out in the update() method and reflected on screen by the view.

Also tweaked some properties: reduced the width of the paddles and slow
the computer AI.