Updates to the GameStateManager and Game Options screen to improve navigation and UI
Updates to the GameStateManager to allow for naviagtion to previous
screens. Introduction of unit tests for this class.

Improvements to the Game Options UI - it is now split into multiple
JPanels to allow for easier layout of components and easier
expandability.

Small changes to other classes such as comments, re-rendering of the
view on change and fixing an issue with hovering of the button not
changing to the original form.