Make use of the new MVC factories
Change the existing controllers and GameStateManager to utilise the MVC
factories. This has the advanatge of reinitialising the compoenents
during each navigation, avoiding setting up multiple action listeners -
bug fixed.
It also means that the GameStateManager has one responsibility which is
to manage the current state, rather that create and initialise the MVC
componenets like it did before.