Link the modifier system to the entities
Work to link the modifier system to the entities in the game world.
When a ball intersects a modifier, it will attach the modifier to the owner enitity (the paddle to last hit the ball) and then set its active status to false (so that the PongModel can cleanup on the next tick).