Move to double precision
Move the entities to double precision. The entities will be rendered using x and y co-ordinates in double precision. Also changed speed to use doubles which will lead into new work with calculating ball collision, angle and speed. This will also help with potential balancing and range of speeds allowed as we now have a larger range to experiment with.