Initial environments implementation
A working stage towards implementing different environments into the game. Each environment has it's own properties that act on the Ball, changing its speed.
The class EnvironmentBall is a new entity in the game world that when hit will change the game environment.
New graphics for Space, Ice and Desert environments have been added to support the above functionality.
Note: intervals for the environment ball spawning will be added in an upcomming commit.