Paddle height fix
Somewhat of a height modification fix. When the change to draw images using double precision was implemented, a bug was introduced where by the modified height wasn't being represented in the view. This was because of using AffineTransforms which only render the image to the actual scale. We therefore need to scale the image when the view detects a change in height from the original.
Note: while this is somewhat of a cumbersome way, it serves the purpose. There may be a way of reconstructing the image, which may be looked at in the future, however this may be too costly at runtime. TBC.