Game music and sound effects
A framework to enable the playing of game music and sound effects on events.
The framework uses AspectJ and custom annotation to invoke sound commands when a method is called. A new class, SoundManager, manages the playing and caching of music and sound files.
ResourceLoader/Manager modified to load and hold sound files on startup.