Fix issue with sounds playing in Jar file
An issue arose with playing sound files inside of a Jar. This was because the InputStream did not have access to mark/reset (read) the stream. Changed InputStream to BufferedInputStream to fix issue.

Also changed the order in which 'soundOverMusic' commands enact, so that the music is stopped first and then the sound to play is started.