Change game to run on 90 second timer and show leaderboard on game over
Changed the game to run on a 90 second timer rather than the max points system.

Added the leaderboard to the game over screen so the user can see the high scores on game end.

With this commit (and the prior 3). i've started to introduce region comments to help with class organisation. This will be applied to each class as it is modified.