Refactor PongView into separate UI components
Refactor the PongView class into separate logical UI components.
This helps separate out responsibility, with GameUIPanel being the parent panel and delegating the UI creation and updating responsibility to the GameInfoPanel and ModifiersPanel.

Note: This work is being done in preparation of moving modifiers to the top of the screen, and making them consumable by players.