Consumable modifiers - WIP
Major step forward in consumable modifiers - major refactor of a lot of components.

Added a number of new modifiers and a new concept of 'Offensive' modifiers to use against opponent. These decrease a property of the opponent on use.

The Computer class has been deleted, with parts of its input functionality moved to InputComponent classes. These will be used on the Player class, with the input types determining whether it is controlled by the user of the computer.

TODO: Still need to give the AI the ability to choose modifiers, fix the decrease modifier issues and add in the reminaing modifier graphics.