Implementation of a player username
When entering a game the user is able to enter a new for themseleves that will then be used in the game itself. This value is cached in preferences and loaded each time a new game is started (as to avoid re-typing). The name is then used in the game display and leaderboards. 
Note that the player name is limited to 20 characters.

Added new layout properties to the Game Over screen including colours for showing who won the game and a line break between scores and leaderboard display.

Consolidated the preferences keys into the Constants class.