Implementation of player colour picker
Added the ability for the user to choose his/her colour in the game options. It will open a JColorChooser dialog (Note: attempted to create custom chooser but ran into issue with colourizing and events - default meets requirements so sticking with that).
This will modify the paddle colour during the game.